Split admin and regular users. Allow delete user
