+++ cmd/server/assets/admin/_nav.html
@@ -27,7 +27,11 @@
   </li>
 
   <li class="nav-item">
-    <a class="nav-link{{if .currentPath.IsDir "/admin/users"}} active{{end}}" href="/admin/users">System admins</a>
+    <a class="nav-link{{if .currentPath.IsDir "/admin/superusers"}} active{{end}}" href="/admin/superusers">System admins</a>
+  </li>
+
+  <li class="nav-item">
+    <a class="nav-link{{if .currentPath.IsDir "/admin/users"}} active{{end}}" href="/admin/users">Users</a>
   </li>
 
   <li class="nav-item">
