+++ cmd/server/assets/admin/superusers/index.html
@@ -1,7 +1,7 @@
-{{define "admin/users/index"}}
+{{define "admin/superusers/index"}}
 
 {{$currentUser := .currentUser}}
-{{$users := .users}}
+{{$admins := .admins}}
 
 <!doctype html>
 <html lang="en">
@@ -16,8 +16,14 @@
     {{template "flash" .}}
 
     <div class="card mb-3 shadow-sm">
-      <div class="card-header">Users</div>
+      <div class="card-header">System admins</div>
       <div class="card-body">
+        <div class="clearfix mb-3">
+          <a class="btn btn-outline-secondary btn-sm float-right" href="/admin/superusers/new">
+            <span class="oi oi-plus" aria-hidden="true"></span>
+            New system admin
+          </a>
+        </div>
 
         <div class="table-responsive">
           <table class="table table-bordered table-striped mb-0">
@@ -29,19 +35,19 @@
               </tr>
             </thead>
             <tbody>
-            {{range $users}}
+            {{range $admins}}
               <tr>
                 <td>{{.Name}}</td>
                 <td>{{.Email}}</td>
                 <td class="text-center">
                   {{- /* cannot delete yourself */ -}}
                   {{if not (eq .ID $currentUser.ID)}}
-                  <a href="/admin/users/{{.ID}}"
+                  <a href="/admin/superusers/{{.ID}}"
                     class="d-block text-danger"
                     data-method="DELETE"
-                    data-confirm="Are you sure you want to delete this user? This will prevent access to all realms system-wide."
+                    data-confirm="Are you sure you want to remove this system admin?"
                     data-toggle="tooltip"
-                    title="Delete this user">
+                    title="Remove this system admin">
                     <span class="oi oi-trash" aria-hidden="true"></span>
                   </a>
                   {{end}}
