+++ internal/routes/server.go
@@ -349,10 +349,13 @@ func Server(
 		adminSub.Handle("/realms/{id:[0-9]+}/leave", adminController.HandleRealmsLeave()).Methods("PATCH")
 		adminSub.Handle("/realms/{id:[0-9]+}", adminController.HandleRealmsUpdate()).Methods("PATCH")
 
+		adminSub.Handle("/superusers", adminController.HandleSuperUsersIndex()).Methods("GET")
+		adminSub.Handle("/superusers", adminController.HandleSuperUsersCreate()).Methods("POST")
+		adminSub.Handle("/superusers/new", adminController.HandleSuperUsersCreate()).Methods("GET")
+		adminSub.Handle("/superusers/{id:[0-9]+}", adminController.HandleSuperUsersDelete()).Methods("DELETE")
+
 		adminSub.Handle("/users", adminController.HandleUsersIndex()).Methods("GET")
-		adminSub.Handle("/users", adminController.HandleUsersCreate()).Methods("POST")
-		adminSub.Handle("/users/new", adminController.HandleUsersCreate()).Methods("GET")
-		adminSub.Handle("/users/{id:[0-9]+}", adminController.HandleUsersDelete()).Methods("DELETE")
+		adminSub.Handle("/users/{id:[0-9]+}", adminController.HandleUserDelete()).Methods("DELETE")
 
 		adminSub.Handle("/mobileapps", adminController.HandleMobileAppsShow()).Methods("GET")
 		adminSub.Handle("/sms", adminController.HandleSMSUpdate()).Methods("GET", "POST")
